payment api fix some bugs

+++ src/app/modules/payment/payment.controller.ts
@@ -1,9 +1,10 @@
 import { RequestHandler } from 'express';
 import booking from '../booking/booking.model';
 import slot from '../slots/slot.model';
+import config from '../../config';
 
 const confirmPayment: RequestHandler = (req, res) => {
-  res.send('<h1>Payment successfull</h1>');
+  res.redirect(`${config.url}/payment/success`);
 };
 
 const failPayment: RequestHandler = async (req, res, next) => {
@@ -13,12 +14,32 @@ const failPayment: RequestHandler = async (req, res, next) => {
 
     await slot.findByIdAndUpdate(findBooking?.slot, { isBooked: 'available' });
 
-    await booking.findByIdAndDelete(id)
+    await booking.findByIdAndDelete(id);
 
-    res.send('<h1>Payment fail</h1>');
+    res.redirect(`${config.url}/payment/fail`);
   } catch (err) {
     next(err);
   }
 };
 
-export const paymentController = { confirmPayment, failPayment };
+const cancelPayment: RequestHandler = async (req, res, next) => {
+  const id = req.params?.id;
+  try {
+    const findBooking = await booking.findById(id);
+
+    if (!findBooking) {
+      res.json({ success: true });
+      return;
+    }
+
+    await slot.findByIdAndUpdate(findBooking?.slot, { isBooked: 'available' });
+
+    await booking.findByIdAndDelete(id);
+
+    res.json({ success: true });
+  } catch (err) {
+    next(err);
+  }
+};
+
+export const paymentController = { confirmPayment, failPayment, cancelPayment };
